adapt all gaudi-c tests to new structure
added  them back to CI
